requirements.txt
requests
feedparser
elasticsearch
celery
redis
python-dotenv
